requirements.txt
flask
flask-cors
pillow
torch
torchvision